refactor : 디렉토리 재구조화
