+++ src/main/java/gift/controller/UserController.java
@@ -1,7 +1,7 @@
 package gift.controller;
 
 import gift.model.user.User;
-import gift.model.user.UserRequest;
+import gift.dto.user.UserRequest;
 import gift.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
