+++ src/main/java/gift/dto/user/UserRequest.java
@@ -1,4 +1,4 @@
-package gift.model.user;
+package gift.dto.user;
 
 public class UserRequest {
