+++ src/main/java/gift/dto/wish/WishResponse.java
@@ -1,4 +1,4 @@
-package gift.model.wish;
+package gift.dto.wish;
 
 public class WishResponse {
