+++ src/test/java/gift/IntegrationTest.java
@@ -1,6 +1,6 @@
 package gift;
 
-import gift.model.user.UserRequest;
+import gift.dto.user.UserRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
